+++ packages/dimensional/src/prefix.ts
@@ -1,3 +1,5 @@
+import { config } from './config';
+
 /**
  * Represents a unit prefix, which is effectively a unit modifier by a scaling factor.
  */
@@ -8,7 +10,7 @@ export class Prefix {
      * @param scale The scale of this prefix, relative to the base unit
      */
     constructor(public readonly LaTeX: string, public readonly scale: number) {
-        if (/^[a-zA-Z]+$/.test(LaTeX)) {
+        if (/^[a-zA-Z]+$/.test(LaTeX) && config.convertToText) {
             this.LaTeX = '\\text{' + LaTeX + '}';
         } else {
             this.LaTeX = '{' + LaTeX + '}';