+++ packages/dimensional/src/quantity.ts
@@ -1,4 +1,6 @@
+import { config } from './config';
 import { Unit } from './unit';
+import * as SMath from 'smath';
 
 /**
  * Represents a physical, measurable quantity.
@@ -71,6 +73,29 @@ export class Quantity {
      * @returns LaTeX code for this quantity
      */
     public toString(): string {
-        return this.quantity.toString() + ' \\left[ ' + this.units.toString() + ' \\right]';
+        const magnitude: number = Math.floor(Math.log10(this.quantity));
+        // Regular notation
+        let quantityString: string = SMath.round2(this.quantity, 10 ** -config.decimalsShown).toString();
+        // Scientific notation
+        if (!SMath.approx(this.quantity, 0, 1e-10) && magnitude !== SMath.clamp(magnitude, config.scientificNotationMagnitude.min, config.scientificNotationMagnitude.max)) {
+            const scaledQuantity: number = this.quantity / (10 ** magnitude);
+            quantityString = SMath.round2(scaledQuantity, 10 ** -config.decimalsShown).toString() + ' \\times 10^{' + magnitude + '}';
+        }
+        // Infinity (+/-)
+        if (this.quantity === Infinity) {
+            quantityString = '\\infty';
+        } else if (this.quantity === -Infinity) {
+            quantityString = '-\\infty';
+        }
+        // NaN
+        if (isNaN(this.quantity)) {
+            quantityString = '\\text{NaN}';
+        }
+        // Unitless
+        if (this.units.is(new Unit()) && !config.showUnitless) {
+            return quantityString;
+        }
+        // Quantity & Unit
+        return quantityString + ' ' + config.unitDelimiters.left + ' ' + this.units.toString() + ' ' + config.unitDelimiters.right;
     }
 }