+++ packages/dimensional/src/test.ts
@@ -1,6 +1,6 @@
 import * as T6 from 't6';
 import * as SMath from 'smath';
-import { prefixes, dimensions, units, Prefix, Dimension, Unit, Quantity } from '.';
+import { prefixes, dimensions, units, Prefix, Dimension, Unit, Quantity, config } from '.';
 import { AmountOfSubstance } from './defaults/dimensions';
 
 {
@@ -31,6 +31,7 @@ import { AmountOfSubstance } from './defaults/dimensions';
     T6.is(units.mile.toString(), '\\text{mi}');
     T6.is(units.Rankine.toString(), '{^{\\circ}\\text{R}}');
     T6.is(units.kilometer.toString(), '{\\text{k}\\text{m}}');
+    T6.is(units.fluidOunce.toString(), '{\\text{fl}_\\text{oz}}');
     T6.is(units.ohm.toString(), units.ohm.over(new Unit()).toString());
     // .dimensions
     T6.isTrue(units.kelvin.dimensions.is(dimensions.Temperature));
@@ -76,6 +77,8 @@ import { AmountOfSubstance } from './defaults/dimensions';
     f = units.foot.to(units.inch);
     T6.eq(f, 12);
     f = units.Rankine.to(units.Rankine);
+    T6.eq(f, 1);
+    f = units.watt.to(units.volt.times(units.ampere));
     T6.eq(f, 1);
     f = units.slug.to(units.poundMass);
     g = units.Gs.to(units.foot.over(units.second.pow(2)));
@@ -184,6 +187,20 @@ import { AmountOfSubstance } from './defaults/dimensions';
     T6.lt(mileTime.quantity, 7.2);
     T6.ge(speedMPH.quantity, 8.3);
     T6.lt(speedMPH.quantity, 8.4);
+    // Scientific Notation
+    const ft_1 = new Quantity(1, units.foot),
+        nm_1 = ft_1.as(units.meter.prefix(prefixes.nano));
+    T6.is(nm_1.toString(), '3.048 \\times 10^{8} \\left[ {\\text{n}\\text{m}} \\right]');
+    const nm_2 = new Quantity(1, units.meter.prefix(prefixes.nano)),
+        ft_2 = nm_2.as(units.foot);
+    T6.is(ft_2.toString(), '3.281 \\times 10^{-9} \\left[ \\text{ft} \\right]');
+    // Infinity
+    const inf_1 = new Quantity(1 / 0, units.ohm);
+    T6.is(inf_1.toString(), '\\infty \\left[ {\\Omega} \\right]');
+    const inf_2 = new Quantity(-1 / 0, units.hertz);
+    T6.is(inf_2.toString(), '-\\infty \\left[ \\text{Hz} \\right]');
+    const nan_1 = new Quantity(-NaN, units.year);
+    T6.is(nan_1.toString(), '\\text{NaN} \\left[ \\text{yr} \\right]');
 }
 
 {
@@ -198,4 +215,31 @@ import { AmountOfSubstance } from './defaults/dimensions';
     T6.lt(height2.quantity, 0.020);
     T6.isTrue(height2.units.is(footballField));
     T6.is(height2.units.toString(), '\\text{fbf}');
+    // Unitless
+    const num = new Quantity(5, units.Unitless);
+    T6.is(num.toString(), '5 \\left[ 1 \\right]');
+    config.showUnitless = false;
+    T6.is(num.toString(), '5');
+}
+
+{
+    // Configuration
+    config.convertToText = false;
+    config.decimalsShown = 1;
+    config.multiplySymbol = '*';
+    config.scalarSymbol = '\\pi';
+    config.unitDelimiters = {
+        left: '[',
+        right: ']',
+    };
+    const digital: Dimension = new Dimension('d');
+    const bit: Unit = new Unit('b', digital);
+    const kilobit: Unit = bit.prefix(prefixes.kilo);
+    const qkb: Quantity = new Quantity(8.88, kilobit);
+    T6.is(digital.toString(), '{d}');
+    T6.is(bit.toString(), '{b}');
+    T6.is(kilobit.toString(), '{\\text{k}{b}}');
+    T6.is(qkb.toString(), '8.9 [ {\\text{k}{b}} ]');
+    T6.is(dimensions.Dimensionless.toString(), '\\pi');
+    T6.is(units.Unitless.toString(), '\\pi');
 }